Reworked string
Removed the "%d (%s)" from the translated part of the string. These variables are the total number of streaming clients and their names, e.g. "3 (John, Adam, Steve)", that is nothing in need of translation. All po files are updated as necessary.

This fixes the following warning when updating the po files:

warning: 'msgid' format string with unnamed arguments cannot be properly localized:
The translator cannot reorder the arguments.
Please consider using a format string with named arguments,
and a mapping instead of a tuple for the arguments.