Directories: add in resolve list only directories that exist (#3366)
* Directories: add in resolve list only directories that exist
* Directories: use one line for loop in addInList

This should increase search speed.
Usually config directories are added at the beginning of the resolve list, which in most cases do not exist for the standard user.
Using resolveFilename these non-existent paths are checked all the time.
This adds in the resolve list only existing directories so as not to check the many non-existent paths on each resolveFilename call.