Fix scaling factor in EpgList
If scaling factor *f is used in setColWidths it raises ValueError because *f cannot be converted as int.
Therefore create a global f = getSkinFactor() in applySkin for this case and use eval() to calculate the scaling factor.
Do not hide skin parameter errors in EpgList.
Print errors in log so that skin or code can be corrected if necessary.