Add new PreviousNext action map
This new action map will allow code to grab or reassign the PREV and NEXT buttons for other uses.

The original button usage design used these buttons for navigation but many users complained and wanted them reserved for editing.  This commit is a workaround to allow these buttons to be remapped back to navigation on a case by case basis.