Add option to add an optional screen via the skin instantly to InfoBar
With long+OK you could toggle between a normal and simple second infobar
but with this option you can when you do not see the infobar at all add
a screen with additional info to the normal infobar and disable it again
the same way. A skinner should be able to add a window with extra info.

example screen in skin (just draws a window with TESTTESTTEST fixed
lable, but it could be SNR bars, resolution and more...

<screen name="InfoBarAdds" title="InfoBarAdd" position="200,150" size="420,64">
	<eLabel position="0,0" size="420,64" font="Bold; 36" halign="left" valign="center" foregroundColor="yellow" backgroundColor="background2" transparent="1" text="TESTTESTTEST" />
</screen>